edited pkgbuild

+++ PKGBUILD
@@ -1,5 +1,8 @@
 # Maintainer: Evgeniy "arcanis" Alexeev <arcanis.arch at gmail dot com>
 
+_build_gui=0
+_build_plasmoid=1
+
 pkgname=kdeplasma-applets-netctl
 _pkgname=netctl-plasmoid
 pkgver=1.0.0
@@ -8,20 +11,22 @@ pkgdesc="Plasmoid written on C++ which interacts with netctl"
 arch=('i686' 'x86_64')
 url="http://arcan1s.github.io/projects/netctlplasmoid"
 license=('GPLv3')
-depends=('kdebase-workspace')
+depends=('netctl' 'qt4')
 makedepends=('cmake' 'automoc4')
+optdepends=('kdebase-runtime: sudo support'
+            'sudo: sudo support')
 source=(https://github.com/arcan1s/netctlplasmoid/releases/download/V.${pkgver}/${_pkgname}-${pkgver}-src.tar.xz)
 install=${pkgname}.install
 md5sums=('1faefe9a5491837863b08fcce79accfe')
-_cmakekeys="-DCMAKE_INSTALL_PREFIX=$(kde4-config --prefix)
-            -DBUILD_GUI:BOOL=1
-            -DBUILD_PLASMOID:BOOL=1
-            -DCMAKE_BUILD_TYPE=Release"
+_cmakekeys="-DCMAKE_INSTALL_PREFIX=$(kde4-config --prefix) -DCMAKE_BUILD_TYPE=Release"
 
 prepare() {
-  if [[ -d ${srcdir}/build ]]; then
-    rm -rf "${srcdir}/build"
-  fi
+  # flags
+  [[ ${_build_gui} == 0 ]] && _cmakekeys=${_cmakekeys}" -DBUILD_GUI:BOOL=0"
+  [[ ${_build_plasmoid} == 1 ]] && depends+=('kdebase-workspace') || _cmakekeys=${_cmakekeys}" -DBUILD_PLASMOID:BOOL=0"
+  
+  # build directory
+  [[ -d ${srcdir}/build ]] && rm -rf "${srcdir}/build"
   mkdir "${srcdir}/build"
 }
